+++ main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"log"
 	"os"
 
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/cli"
@@ -10,7 +9,5 @@ import (
 var version = ""
 
 func main() {
-	if err := cli.Init(&version).Run(os.Args); err != nil {
-		log.Fatal(err)
-	}
+	cli.Init(&version).Run(os.Args)
 }